requirements.txt
PyYAML>=5.3.1
GitPython>=3.1.
alive-progess>=3.1.5
PyGithub>=1.47
